+++ apps/snappyHexMesh/view/config/content/4_Refinement/RefinementProps/Surface.qml
@@ -27,7 +27,7 @@ Column {
         spacing: 10
         enabled: convertToRegion.checked
         visible: enabled
-//        height: enabled ? childrenRect.height : 0
+        height: enabled ? childrenRect.height : 0
         width: parent.width
 
         Subheader {